requirements.txt
addict==2.4.0
albumentations==1.4.21
basicsr==1.4.2
cityscapesscripts==2.2.4
clip==0.2.0
easydict==1.13
einops==0.8.0
kornia==0.6.12
lmdb==1.5.1
matplotlib==3.9.2
mc==0.0
numpy==1.21.6
omegaconf==2.3.0
onnx==1.17.0
open_clip_torch==2.0.2
packaging==24.1
parrots==1.0.3
Pillow==9.5.0
Pillow==11.0.0
prettytable==3.12.0
pytorch_lightning==1.5.0
PyYAML==6.0.1
PyYAML==6.0.2
regex==2023.12.25
safetensors==0.4.5
scikit_image==0.19.3
scipy==1.7.3
tensorboardX==2.6.2.2
tifffile==2021.11.2
timm==0.9.12
torch==1.13.1
torch_tb_profiler==0.4.3
torchvision==0.14.1
tqdm==4.66.2
transformers==4.30.2
turbojpeg==0.0.2
xformers==0.0.28.post3
yapf==0.40.2
